requirements.txt
funasr
fastapi
uvicorn
python-multipart
numpy<=1.26.4
ffmpeg-python